Reimplement generateFeatureSet
Reimplemented method:
* uses a list of Value instances as its only parameter
* returns a map of parameter id to a selected parameter value id
* selects values for parameters in order from parameter whose
value is know for the most languages to parameter with
least coverage in source data - this will ensure the
part of selection algorithm relying on probability values
will be used as long as possible
* uses probability independent random selection for parameters with
probablity of co-occurence being 0.0 - which handles cases of two
parameter values not being attested together in a language due
to missing data

The function will undergo further refactoring.
Also, cases in which parameter values never co-occur together because
of one value excluding the other (like SOV-related parameters for
languages displaying SVO word order) must be handled.