Fix Feature.influencingFeatures field
The field didn't work properly - attempting to save
a Feature instance using it ended with an exception
with a message "Table does not reference target".
Internally, Exposed couldn't recognize which reference
to Feature on FeatureInfluencingFeatures should be used.

It turns out relationships between different entities
of the same class should be expressed differently:
https://github.com/JetBrains/Exposed/wiki/DAO#parent-child-reference